Fixed bug preventing creation of new news
New release as V.0.5.1

+++ Ålleinfo Admin/NewsItem.cs
@@ -134,7 +134,7 @@ namespace Ålleinfo_Admin
             {
                 try
                 {
-                    Webber.removeNews(data.id);
+                    Webber.removeNews((int)data.id);
 
                     AdminForm.adminForm.removeNewsDisplay(this);
                 }
